consolidate SageCommand
