+++ src/sage-lib/sage/model/SageReaction.ts
@@ -1,19 +1,19 @@
 import { Cache } from "@rsc-utils/cache-utils";
 import { debug } from "@rsc-utils/console-utils";
 import type { DMessage, DReaction, DUser } from "@rsc-utils/discord-utils";
-import { ReactionType } from "../../discord";
-import type { HasSageCacheCore } from "./HasSageCache";
-import { HasSageCache } from "./HasSageCache";
-import { SageCache } from "./SageCache";
+import { ReactionType } from "../../discord/index.js";
+import { GameRoleType } from "./Game.js";
+import { SageCache } from "./SageCache.js";
+import { SageCommand, type SageCommandCore } from "./SageCommand.js";
 
-interface SageReactionCore extends HasSageCacheCore {
+interface SageReactionCore extends SageCommandCore {
 	messageReaction: DReaction;
 	user: DUser;
 	reactionType: ReactionType;
 };
 
 export class SageReaction
-	extends HasSageCache<SageReactionCore, SageReaction> {
+	extends SageCommand<SageReactionCore, SageReaction> {
 
 	private constructor(protected core: SageReactionCore, cache?: Cache) {
 		super(core, cache);
@@ -21,17 +21,20 @@ export class SageReaction
 	public clear(): void {
 		debug("Clearing SageReaction");
 		this.cache.clear();
-		this.caches.clear();
+		this.sageCache.clear();
 	}
 
 	public static async fromMessageReaction(messageReaction: DReaction, user: DUser, reactionType: ReactionType): Promise<SageReaction> {
-		const caches = await SageCache.fromMessageReaction(messageReaction, user);
-		return new SageReaction({
-			caches,
+		const sageCache = await SageCache.fromMessageReaction(messageReaction, user);
+		const sageReaction = new SageReaction({
+			sageCache,
 			messageReaction,
 			reactionType,
 			user
 		});
+		sageReaction.isGameMaster = await sageReaction.game?.hasUser(user.id, GameRoleType.GameMaster) ?? false;
+		sageReaction.isPlayer = await sageReaction.game?.hasUser(user.id, GameRoleType.Player) ?? false;
+		return sageReaction;
 	}
 	public get isAdd(): boolean {
 		return this.core.reactionType === ReactionType.Add;
